Refactor: Extract get updated fields helper methods
