+++ src/utils/updatedFields/hasUpdate.ts
@@ -1,0 +1,2 @@
+export const hasUpdate = (originalValue?: string | Date | null, updatedValue?: string | Date | null): boolean =>
+  !!updatedValue && updatedValue !== originalValue
